+++ src/components/breadcrumb-navigation/breadcrumb--navigation-arrow.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { ChevronRightIcon } from "../../icons";
 
-export const BreadcrumbsArrow = () => {
+export const BreadcrumbNavigationArrow = () => {
     return <ChevronRightIcon className="h-3 w-3 text-neutral-800" />;
 };
