+++ src/components/breadcrumb-navigation/index.tsx
@@ -1,0 +1,1 @@
+export { BreadcrumbNavigation } from "./breadcrumb-navigation";
