--- src/components/breadcrumbs/index.tsx
@@ -1,1 +1,0 @@
-export { Breadcrumbs } from "./breadcrumbs";
